Changes needed for Visual C++ 2005 support
--HG--
extra : convert_revision : svn%3Aad96d65e-77c6-4fb4-b022-7d5d8cc5d1e0/trunk%4028

+++ app/help/genhelp.c
@@ -68,7 +68,7 @@ void process( mode_e mode, FILE * inFile, FILE * outFile )
 		lineNum++;
 		if (line[0] == '#')
 			continue;
-		len = strlen( line );
+		len = (int)strlen( line );
 		if (line[len-1] == '\n' ) len--;
 		if (line[len-1] == '\r' ) len--;
 		line[len] = '\0';
